refactor: split data functions over more files
