--- src/apiserver/app/routers/auth/__init__.py
@@ -1,3 +1,0 @@
-from apiserver.app.routers.auth.auth import router as auth_router
-
-__all__ = ["auth_router"]
