+++ src/apiserver/app/routers/profile.py
@@ -15,6 +15,6 @@ async def get_profile(request: Request, authorization: Authorization):
     dsrc: Source = request.state.dsrc
     acc = await handle_auth(authorization, dsrc)
     async with data.get_conn(dsrc) as conn:
-        user_data = await data.user.get_userdata_by_id(conn, acc.sub)
+        user_data = await data.ud.get_userdata_by_id(conn, acc.sub)
 
     return user_data
