+++ src/apiserver/data/schema.py
@@ -1,9 +1,10 @@
 from apiserver.data.api.refreshtoken import RefreshOps
-from apiserver.data.api.user import UserOps, UserDataOps
+from apiserver.data.api.user import UserOps
+from apiserver.data.api.ud.userdata import UserDataOps
 from auth.data.schemad.ops import SchemaOps
 
 
-__all__ = ["SCHEMA"]
+__all__ = ["OPS"]
 
 
-SCHEMA = SchemaOps(user=UserOps, userdata=UserDataOps, refresh=RefreshOps)
+OPS = SchemaOps(user=UserOps, userdata=UserDataOps, refresh=RefreshOps)
