+++ src/auth/data/__init__.py
@@ -3,5 +3,6 @@ from auth.data import authentication
 from auth.data import register
 from auth.data import token
 from auth.data import keys
+from auth.data import update
 
-__all__ = ["authorize", "authentication", "register", "token", "keys"]
+__all__ = ["authorize", "authentication", "register", "token", "keys", "update"]
