+++ src/auth/data/schemad/user.py
@@ -15,10 +15,17 @@ class UserOps(Protocol):
     async def get_user_by_email(cls, conn: AsyncConnection, email: str) -> User:
         ...
 
+    @classmethod
+    async def update_password_file(
+        cls, conn: AsyncConnection, user_id: str, password_file: str
+    ) -> int:
+        ...
+
 
 class UserDataOps(Protocol):
     @classmethod
     async def get_userdata_by_id(cls, conn: AsyncConnection, user_id: str) -> UserData:
+        """Throws NoDataError if user does not exist."""
         ...
 
     @classmethod
